pyproject.toml
[tool.poetry]
name = "verschenken-finder"
version = "0.1.0"
description = "This app wil find Verschenken Offers in the Kleineanzeigen website."
authors = ["Lee <as.alisayyadi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.9"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
fastapi = "^0.115.6"
pymongo = "^4.10.1"
requests = "^2.32.3"
pytest = "^8.3.4"
uvicorn = "^0.34.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

package-mode = false